doc: README 에 변경내용 반영
