+++ README.md
@@ -11,9 +11,11 @@
 $> git clone https://github.com/jozoshu/I_WANT_JOB.git
 
 .
-├── modules            # 모듈
-│   ├── crawlers
+├── config             # 환경 모듈
 │   ├── db
+│   └── logging
+├── modules            # 실행 모듈
+│   ├── crawlers
 │   └── handlers
 ├── venv               # python 가상환경
 ├── .env               # 환경변수 파일
@@ -90,5 +92,5 @@ $> source venv/bin/activate
 
 ## 1.3 Run~!
 ~~~bash
-$> python main.py
+(venv) $> python main.py
 ~~~
